django_blog_api/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
